+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Improvements
 
 - [cronos#720](https://github.com/crypto-org-chain/cronos/pull/720) Add option `iavl-disable-fastnode` to disable iavl fastnode indexing migration (backport #714).
+- [cronos#721](https://github.com/crypto-org-chain/cronos/pull/721) Integrate the file state streamer (backport #702).
 
 *Sep 20, 2022*